Nueva rama

+++ src/view/project-detail-page/ProjectDetailPage.jsx
@@ -1,51 +1,73 @@
+import styles from './ProjectDetailPage.module.css'
 import { useState, useEffect } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { useNavigate, useParams } from 'react-router-dom'
+import { ThumbsUp } from 'lucide-react'
+import { Button } from 'react-bootstrap'
+import Toastify from 'toastify-js'
+import 'toastify-js/src/toastify.css'
 import {
 	getProjectById,
 	getUserProfile,
 	deleteProject,
+	toggleProjectLike,
 } from '../../redux/actions'
-import { Button, Modal } from 'react-bootstrap'
-import styles from './ProjectDetailPage.module.css' // Asegúrate de importar el CSS módulo
+import UpdateProjectPage from '../update-project-page/UpdateProjectPage'
 
 const ProjectDetailPage = () => {
+	const dispatch = useDispatch()
 	const navigate = useNavigate()
 	const { id } = useParams()
 	const { project } = useSelector((state) => state.projects)
 	const { token, loggedUser } = useSelector((state) => state.auth)
-	const dispatch = useDispatch()
-
-	const [show, setShow] = useState(false)
+	const { theme } = useSelector((state) => state.themes)
+	const [showEditModal, setShowEditModal] = useState(false)
 
 	useEffect(() => {
 		dispatch(getProjectById(id))
 		dispatch(getUserProfile(token))
 	}, [dispatch, id, token])
 
-	const handleEdit = () => {
-		navigate(`/modProject/${id}`)
+	const handleEdit = () => setShowEditModal(true)
+	const handleClose = () => {
+		setShowEditModal(false)
 	}
 
-	const handleDelete = () => {
-		dispatch(deleteProject(id, token))
-		alert('Proyecto eliminado exitosamente')
-		navigate('/myprofile')
+	const archiveProject = () => {
+		const confirm = window.confirm(
+			'¿Estás seguro que deseas archivar el proyecto?'
+		)
+		if (confirm) {
+			dispatch(deleteProject(id, token))
+			Toastify({
+				text: 'Proyecto archivado',
+				duration: 3000,
+				close: true,
+				gravity: 'top',
+				position: 'center',
+				backgroundColor: '#4CAF50',
+				stopOnFocus: true,
+			}).showToast()
+			navigate('/myprofile')
+		}
 	}
 
-	const handleClose = () => setShow(false)
-	// const handleShow = () => setShow(true)
-
-	const fileProject = () => {
-		dispatch(deleteProject(id, token))
-		alert('Proyecto archivado')
-		navigate('/myprofile')
+	const toggleLike = (project) => {
+		if (loggedUser && project) {
+			dispatch(
+				toggleProjectLike(
+					{ projectId: project.id, userId: loggedUser.id },
+					project.id,
+					token
+				)
+			)
+		}
 	}
 
 	if (!project) return null
 
 	return (
-		<div className="container mt-5">
+		<div className={`styles.container`}>
 			<section className={styles.card}>
 				<div className={styles.cardImageContainer}>
 					<img
@@ -55,20 +77,19 @@ const ProjectDetailPage = () => {
 					/>
 				</div>
 				<div className={styles.cardDetail}>
-					<p className={styles.cardDetailCaption}>PROJECT</p>
-					<h1 className={styles.cardDetailTitle}>{project?.title}</h1>
+					<div className={styles.userProfile}>
+						<img src={project?.user?.image} alt={project?.id} />
+						<span>{project.user.userName}</span>
+					</div>
+					<h2 className={styles.cardDetailTitle}>{project?.title}</h2>
 					<p className={styles.cardDetailDesc}>{project?.description}</p>
 
 					{project?.tags && (
 						<div className={styles.tagsContainer}>
-							{project?.tags.map((tag, index) => (
-								<button
-									key={index}
-									className={`btn ${styles.tagButton}`}
-									disabled
-								>
-									{tag}
-								</button>
+							{project?.tags.map((tag) => (
+								<span key={tag.id} className={styles.tagText}>
+									#{tag.tagName}
+								</span>
 							))}
 						</div>
 					)}
@@ -78,58 +99,52 @@ const ProjectDetailPage = () => {
 					{project?.technologies && (
 						<div className={styles.technologiesContainer}>
 							{project?.technologies.map((tech) => (
-								<button
-									key={tech.id}
-									className={`btn ${styles.techButton}`}
-									disabled
-								>
+								<span key={tech.id} className={styles.techButton}>
 									{tech.name}
-								</button>
+								</span>
 							))}
 						</div>
 					)}
 
-					{loggedUser?.id !== undefined && loggedUser?.id === project?.userId && (
-						<div className={styles.buttonsContainer}>
-							<button
-								className={`btn ${styles.modifyButton}`}
-								onClick={handleEdit}
-								style={{ textDecoration: 'none', margin: '10px' }}
-							>
-								Modificar
-							</button>
-							{/* <button
-								className={`btn ${styles.deleteButton}`}
-								onClick={handleShow}
-								style={{ textDecoration: 'none', margin: '10px' }}
-							>
-								Eliminar
-							</button> */}
-							<Modal show={show} onHide={handleClose}>
-								<Modal.Header closeButton>
-									<Modal.Title>Confirmación</Modal.Title>
-								</Modal.Header>
-								<Modal.Body>
-									¿Estás seguro que deseas eliminar el proyecto?
-								</Modal.Body>
-								<Modal.Footer>
-									<Button variant="secondary" onClick={handleClose}>
-										Cancelar
-									</Button>
-									<Button variant="danger" onClick={handleDelete}>
-										Aceptar
-									</Button>
-								</Modal.Footer>
-							</Modal>
-							<button
-								className={`btn ${styles.fileButton}`}
-								onClick={fileProject}
-								style={{ textDecoration: 'none', margin: '10px' }}
-							>
-								Archivar
-							</button>
-						</div>
-					)}
+					<div
+						onClick={() => toggleLike(project)}
+						className={`${styles.likeButton} mt-4`}
+					>
+						<span>{project.likes.length}</span>
+						<ThumbsUp
+							className={theme === 'light' ? 'text-dark' : 'text-light'}
+							fill={
+								project.likes.some((like) => like?.userId === loggedUser?.id)
+									? theme === 'light'
+										? '#343a40'
+										: '#f8f9fa'
+									: 'none'
+							}
+							size={20}
+						/>
+					</div>
+					{loggedUser?.id !== undefined &&
+						loggedUser?.id === project?.userId && (
+							<div className={styles.buttonsContainer}>
+								<Button variant="primary" onClick={handleEdit}>
+									Modificar
+								</Button>
+								<Button variant="secondary" onClick={archiveProject}>
+									Archivar
+								</Button>
+
+								{showEditModal && (
+									<div className={styles.modalOverlay} onClick={handleClose}>
+										<div
+											className={styles.modalContent}
+											onClick={(e) => e.stopPropagation()}
+										>
+											<UpdateProjectPage />
+										</div>
+									</div>
+								)}
+							</div>
+						)}
 				</div>
 			</section>
 		</div>
